relocate StyledGridWrapper
