+++ components/GlobalStyles.js
@@ -26,6 +26,12 @@ const GlobalStyles = createGlobalStyle`
 
 export default GlobalStyles;
 
+export const StyledGridWrapper = styled.div`
+  display: grid;
+  grid-template-rows: auto 1fr auto;
+  height: 100vh;
+`;
+
 export const StyledMain = styled.main`
   display: flex;
   flex-direction: column;
