+++ pages/createpage/index.js
@@ -1,8 +1,7 @@
 import React from "react";
-import { StyledGridWrapper } from "..";
 import Footer from "../../components/Footer/Footer";
 import CreateMemoryForm from "../../components/Forms/CreateMemoryForm";
-import { StyledMain } from "../../components/GlobalStyles";
+import { StyledGridWrapper, StyledMain } from "../../components/GlobalStyles";
 import Header from "../../components/Header/Header";
 
 export default function CreatePage({
